backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pillow==10.3.0
pytesseract==0.3.10
groq==0.9.0
python-dotenv==1.0.1
aiofiles==23.2.1
pydantic==2.7.4
httpx==0.27.0
pdf2image==1.17.0
sqlalchemy==2.0.31